+++ lib/services/pac_provider_fm.rb
@@ -1,4 +1,4 @@
-require 'pac_provider_service'
+require 'services/pac_provider_service'
 require 'fm_timbrado_cfdi'
 
 class PacProviderFM < PacProviderService
