+++ lib/services/pac_provider_service.rb
